Add an AsyncState to communicate between acquire() and produce()
Add a template type parameter to the edm::EDProducerExternalWork class.

The acquire() method is now const, and it communicates its asynchronous
state to the the produce() method via an object of this type, rather
than using data members.

Implementations that do not need to pass any state between acquire() and
produce() can use "void" or leave the template parameter empty.