Fix the shorten function
If the size of type T is smaller than 8 bytes, the existing shorten
function will modify memory beyond the location of the existing argument.

The fix creates a temporary value of 8 bytes in the function to perform the
operations on.  It still reads too much memory (for T < 8 bytes), but it
will no longer write to that memory.

To be completely memory safe, the unsigned temporary type could be
chosen with std::conditional, like
  using utype = std::conditional<sizeof(float) == 4, uint32_t, uint64_t>::type;
It would need more branches to handle the size 1 and 2 case as well.

The casts to and from uint64_t involve pointers rather than values
because we want to transfer the bit patterns, not the values.