Fix uploading the EventSetup conditions from multiple streams
When multiple CUDA streams are trying to initialise the same EventSetup object, the first one to do
so starts the asynchronous operations, and the others are supposed to wait for it to finish.
However, code for recording the CUDA event was missing, so the other streams would find the default-
constructed event, which is always "valid".

Adding the missing call to record the event fixes the problem.