Workarounds for the HIP port
Fix the radixSort_t test.

Remove the launch bounds, which seem to cause the kernel to fail to run
(and no error is raised)

Add  __threadfence after the update of the ibs variable, which controls
the enclosing while loop.  Without this threadfence, the loop appears to
keep running, and the preceeding assert will trigger (c[bin] >= 0).