[cudadev] Added support for buffer-store-view SoA model.
The SoA store divides a user pre-allocated buffer in adjacent columns of values. The columns are bytes aligned to a setable alignment
and their length is determined at run time.

The SoA views allow grouping columns from multiple stores in a logical entity. For example when some columns are used on device-only
and otheres transmitted to host. The two groups are on two different stores, but joined together in a logic view.

Views can also provide access to a subset of a store.