[alpaka] Added SoA layouts and views and a test
The test validates multiple multiple SoA use cases, host and device side.
It validates buffer sharing for multiple layouts.
An an Alpaka native sub-buffer operation (device side memset) was used. This type of operations
can be used in the future to copy partial columns.
This version is a port of https://github.com/cms-patatrack/pixeltrack-standalone/pull/287 to Alpaka.