[alpaka] Remove the cms::alpakatools::ALPAKA_ACCELERATOR_NAMESPACE
Many functions are templated on the Accelerator types, so they didn't
need to be in the ALPAKA_ACCELERATOR_NAMESPACE.

The device global variables, device list and initialisation functions
are now templated on the Platform type.

The ESProduct class template is now templated on the Queue type.

Simplify the use of the cms::alpakatools namespace.