[alpaka] add option to compile for the SYCL backend
The application is compiled once for the CPU(s) and once for the Intel GPU(s).
The flags for AOT compilation have been added and icpx is used as the default compiler for the SYCL backend.
The linking step is performed as well with icpx when the SYCL backend is enabled.